flask-backend/requirements.txt
# Web framework
Flask==2.3.2

# Database ORM
SQLAlchemy==1.4.41

# SQLite migrations
Flask-Migrate==3.1.0

# Request validation
marshmallow==3.14.1

# CORS handling
Flask-Cors==3.0.10

# Environment variable management
python-dotenv==0.21.0

# WSGI server for production
gunicorn==20.1.0

# Werkzeug (pas besoin de forcer une version spécifique)
# Flask inclut automatiquement la version correcte

# Flask Bcrypt pour le hachage des mots de passe
Flask-Bcrypt==1.0.1

# JWT handling
Flask-JWT-Extended==4.4.4

# If using PostgreSQL, you can add psycopg2-binary
psycopg2-binary>=2.9.3

# Socket.IO support
Flask-SocketIO==5.3.4

gevent>=22.10.2
gevent-websocket>=0.10.1
